% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenew.r
\name{get_me_a_new_password}
\alias{get_me_a_new_password}
\title{Get a new password}
\usage{
get_me_a_new_password(lengthPass = 6, type = "EFF", nInt = 2)
}
\arguments{
\item{lengthPass}{The number of words in your password. Default is 6}

\item{type}{The type of passphrase list to generate. The default is "EFF". You can select from "Harry Potter", "Star Trek", "Game of Thrones", "Star Wars"}

\item{nInt}{The number of integers to include. You can turn them into a special character on your own.}
}
\value{
a new password for you
}
\description{
Get a new password
}
\examples{
get_me_a_new_password(5, "Harry Potter")
}
